modified integer variables throughout dbSeed.sql to include character limit parameter

+++ sql/dbSeed.sql
@@ -17,7 +17,7 @@ CREATE TABLE role (
     id INTEGER(10) NOT NULL AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(30) NOT NULL,
     salary DECIMAL(10,2) NOT NULL,
-    deparment_id INTEGER,
+    deparment_id INTEGER (10),
     FOREIGN KEY (department_id) REFERENCES department(id)
 );
 
@@ -26,8 +26,8 @@ CREATE TABLE employee (
     id INTEGER(10) NOT NULL AUTO_INCREMENT PRIMARY KEY,
     first_name VARCHAR(30) NOT NULL,
     last_name VARCHAR(30) NOT NULL,
-    role_id INTEGER,
+    role_id INTEGER (10),
     FOREIGN KEY (role_id) REFERENCES role(id),
-    manager_id INTEGER,
+    manager_id INTEGER(10),
     FOREIGN KEY (manager_id) REFERENCES employee(id)
 );
